Set the clk_flags field to default configuration in the i2c configuration.
The struct i2c_config_t now has a new field called clk_flags in the most
recent stable version of the ESP-IDF,  ESP-IDF v4.4.2-dirty When self-balancing
and MPU code are flashed, the i2c configuration fails because the field is not
initialized in the present code, which results in the ESP32 rebooting.

Set the default value for the clk_flags in the i2c configuration.