MPU calibration
When the bot is in a steady horizontal position the value recorded by the gyroscope and accelerometer of the MPU6050 should be :
gx = 0 gy = 0 gz = 0 and ax = 0 ay = 0 az = 16384

But, we get some error in the values at the steady horizontal position.
This PR aims to add a function to calculate optimum offsets for MPU6050.
Subtracting these offsets from accelerometer and gyroscope raw values will account for the zero error caused.

The approach followed is as follows:
1. Keep the bot in a horizontal stable state when the flashed code starts running.
2. Take an average of 1000 readings and set the offset to these readings.
3. Take the average of next 1000 readings, but deduct the offset obtained in step 2 from each raw reading.
4. Add this average to the offset.
5. Repeat this process until the average becomes close to zero.